Try appending the vim format string
livemd files need this vim format string to render properly on GitHub.
Unfortunately LiveBook 0.4 overwrites the livemd files when it saves and
does not include the vim format string. That means I have to ensure that
formatting is added when I commit.

If I can append the format string then that's easy

```
$ comm -23 <(fd livemd) <(rg -l '<!-- vim: syntax=markdown -->') | while read f; do echo '<!-- vim: syntax=markdown -->' >>| "$f"; done
```