Advent of Code 2021 - Day 14
Whew this was a tough one for me to conceptualize until it finally
clicked. We don't need to transform and construct the actual polymer
string but turn the initial string into a set of pairs and continually
transform the pairs per the rules. At the end we only need to count the
left character of each pair because of how we split the string. We
ensure that each pair has two characters and if it doesn't (i.e. the
trailing pair) we insert [0] simply as a placeholder.

The rules application output is a list of pairs and their frequencies

```elixir
%{
  [66, 0] => 1,
  'BB' => 9,
  'BC' => 4,
  'BH' => 3,
  'BN' => 6,
  'CB' => 5,
  'CC' => 2,
  'CN' => 3,
  'HC' => 3,
  'HH' => 1,
  'HN' => 1,
  'NB' => 9,
  'NC' => 1,
  'NH' => 1
}
```

We iterate that list, consider only the left character of the pair, and
add the count of that pair to the total for that left character.

And boom! Almost instantly processed polymer vs an infinity of building
lists with trillions of elements.