Advent of Code 2021 - Day 15
Wowzers I thought I was losing it on this one. Turns out the libgraph
library had an actual problem with large number of vertices because it
was using a hashing algorithm that caused collisions.

The result was that I had 250000 cells but when transformed into a graph
it would only have ~249993 vertices. That is about 7 vertices would
collide and not be added to the graph.

The fix is to run the unreleased version of libgraph that allows
declaring a custom function to use to hash the vertices. In my case I
could simply use the coordinates as the vertex ids.